refactor: extract services following SOLID principles
